Split document rendering from rendering of the tree
This allows us to stop passing the document around, by asserting
that not only must the document be the top-level element, but also
that it can only contain header, content, and footer elements. This
breaks consumers of the API in that they now need to explicitly
include content at the top level as well as header and footer.